requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
huggingface-hub>=0.19.0
pypdf>=3.17.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
